@@ -1,16 +1,17 @@
-import {Injectable} from '@angular/core';
-import {XHRHandler} from './xhrhandler.service';
+import {Injectable}   from '@angular/core';
+import {XHRHandler}   from './xhrhandler.service';
 import {LoginRequest} from '../data/loginRequest'
+import {Observable}   from 'rxjs';
 
 @Injectable() 
 export class AppService {
   constructor(private xhrhandler: XHRHandler) {}
 
-  userLogin(request: LoginRequest): Promise<any> {
+  userLogin(request: LoginRequest): Observable<any> {
     return this.xhrhandler.doPost('user/login', request);
   }
 
-  listUser(): Promise<any> {
+  listUser(): Observable<any> {
     return this.xhrhandler.doGet('user/list');
   }
 }
